pilldispenser-rpi/requirements.txt
#Requirements
uvicorn[standard]
fastapi
gpio.rpi
requests